% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_female_words.r
\name{count_female_words}
\alias{count_female_words}
\title{A Function to count the female words per episode and save it in a dataframe}
\usage{
count_female_words(i, j, data_scgen, data_wpe)
}
\description{
This function tracks the absolute number of words spoken by women. It first adds up all words and finally stores the final amount of female and male words of every episode in a new dataset. However, as the function stands now the ordering did not work very well so that it creates an incomplete summary of word counts.
}
\examples{
count_female_words()
}

