% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.r
\name{load_data}
\alias{load_data}
\title{A Function to load the datasets.}
\usage{
load_data(data_script, data_gender, data_imdb)
}
\description{
This function loads the three datasets you download from kaggle: https://www.kaggle.com/wcukierski/the-simpsons-by-the-data
}
\examples{
load_data()
}

