% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_and_clean.r
\name{merge_and_clean}
\alias{merge_and_clean}
\title{A function to merge and clean unimportant variables and observations}
\usage{
merge_and_clean(data_gender, data_script)
}
\description{
This function clears all unimportant oberservations and variables and merges the two important sets data_script and data_gender
}
\examples{
merge_and_clean()
}

